Histogram classes reorganized
