+++ test/unit_tests/core/observables/correlation/DensityHistogramTest.cpp
@@ -41,9 +41,9 @@ TEST_CASE("DensityHistogram") {
     }
 
     auto histogram = densityHistogram.dumpValues();
-    auto zerosRemover = [](const BinValue<3> &binValue) { return binValue.value == 0; };
+    auto zerosRemover = [](const Histogram3D::BinValue &binValue) { return binValue.value == 0; };
     histogram.erase(std::remove_if(histogram.begin(), histogram.end(), zerosRemover), histogram.end());
-    std::vector<BinValue<3>> expected = {
+    std::vector<Histogram3D::BinValue> expected = {
         {{0.45, 0.5, 0.15}, 12.5}, {{0.45, 0.5, 0.75}, 12.5},
         {{0.55, 0.5, 0.15}, 25}, {{0.55, 0.5, 0.75}, 25},
         {{0.65, 0.5, 0.15}, 12.5}, {{0.65, 0.5, 0.75}, 12.5}
